Make SonarCloud happy.
The removal of `<?>` is requested by SonarCloid to make it happy but doing this is a bad practice.
This is not a big issue so just add the SuppressWarnings for rawtypes to appease SonarCloud 's bias.

The `@Autowire` will not be removed as per recommendation by SonarCloud as doing so will make the code significantly less maintainable.
If there is a bug in how `@Autuowire` operates, then that should be an issue for Spring and not as a "Code Smell" in SonarCloud.
Manually disable SonarClouds bad practice in this case.

In the case of `HttpService.class`, the history of the file shows that the rest builder had to be manually allocated as is done in that code.
This particular case of SonarCloud complaints is likely bad advice and following such may introduce bugs that was fixed by this particular setup using `@PostConstruct`.