Add Junit Jupiter and Mockito for testing and begin adding some tests.
The controller tests should be MVC tests.
- Move the `TenantControllerTest` test code into the `HibernateSchemaServiceIT` integration test.
- Add and use MockMVC for the `TenantControllerTest`.
- Use parameterized testing in `TenantControllerTest` to handle a common set of other protocols.

I wanted to write unit tests for `HibernateSchemaService` but the direct use of `TenantProperties` (and similar) overly complicates the matter.
Continue to use an integration test, but name the test source file with `IT` at the end to designate this.

Add the missing `hikari` settings to the `application.yml` file.

There is a fixme comment added regarding the behavior of when no okapi headers are provide.
It may be that the code is allowing a later part to fail than at the immediate controller level.
In which case, the mocked createTenant() is preventing this failure case.
Additional investigation on this is needed.

There is a fixme comment about the Content-Type being returned that is not in the Accept header.
This endpoint should probably be returning a 415.
The intended behavior is unclear and so the test is written to pass based on the current behavior.