Add spring-test project, add utility to spring-web, and update spring-domain to use spring-test.
Add spring-test project for helping assist with test writing.
Update the spring-domain to utilize this.
The spring-tenant needs to get updated (but I ran out of time).

This adds and implements the HTTP Accept header validations and provides a validation response.
As far as I am able to find (thus far), Spring-Boot does not provide any functionality to automatically validate that the Accept header is within an accepted range.
Add a helper utility to spring-web and use this to help validate the Accept header and return a 415 when this Accept header is not acceptable.